requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.0.3
tensorflow==2.15.0
scapy==2.5.0
pyshark==0.6.0
Flask==3.0.2
matplotlib==3.8.2
plotly==5.18.0
mysql-connector-python==8.3.0
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-Caching==2.1.0
python-socketio==5.10.0
